requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
pandas
numpy
scikit-learn
google-genai
pandapower
networkx
matplotlib
requests
